requirements.txt
pandas
numpy
scikit-learn
streamlit
google-cloud-bigquery
replicate
joblib